+++ CHANGELOG.md
@@ -13,7 +13,7 @@ CHANGELOG
 - Fix bug where Drop-in would not finish loading if inside a hidden div
 - Improve transition from payment sheet views to payment methods view
 - Use version 3.19.1 of braintree-web
-- Add `isPaymentMethodSelected` property on `paymentMethodRequestable` events
+- Add `paymentMethodIsSelected` property on `paymentMethodRequestable` events
 - Improve UI in older versions of iOS Safari
 
 1.3.1